education/pyproject.toml
[tool.poetry]
name = "education"
version = "0.1.0"
description = ""
authors = ["ramazan_aisaev <aisaev-68@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.5"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
python-decouple = "^3.8"
docutils = "^0.20.1"
django-cors-headers = "^4.2.0"
mimesis = "^11.1.0"
faker = "^19.6.2"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
